V2 - MD5 Working + Folders Fixed
The Md5 Check now deletes the folder if any images within it have changed to ensure that the images that were there are removed first. This fixes if less images are exported and it tries to update, meaning the extra photos would remain. It now also looks for a folder called Upload within the Export folder, as these are the ones to  be backed up / uploaded / kept. It will just copy ALL jpg's if this folder doesnt exist.